daemon/requirements.txt
PyMySQL==1.0.2
cryptography==3.4.7
Flask==1.1.2
flask_sqlalchemy==2.5.1
sqlalchemy < 1.4.0
flask_migrate==2.7.0
flask_script==2.0.6
sqlalchemy_utils==0.36.8
flask-jwt-extended==4.1.0
redis==3.5.3
requests==2.25.1
pyjwt==2.1.0
python-dateutil==2.8.1
pytz